move and re-use TorchProfiler class
