+++ openpifpaf/decoder/__init__.py
@@ -5,6 +5,5 @@ from .decoder import Decoder
 from .cifcaf import CifCaf
 from .cifdet import CifDet
 from .factory import cli, configure, factory
-from .profiler import Profiler, TorchProfiler
 
 from .factory import DECODERS
