+++ openpifpaf/decoder/factory.py
@@ -7,7 +7,7 @@ from .cifdet import CifDet
 from .decoder import Decoder
 from .multi import Multi
 from . import utils
-from .profiler import Profiler, TorchProfiler
+from ..profiler import Profiler, TorchProfiler
 
 LOG = logging.getLogger(__name__)
 
@@ -155,7 +155,7 @@ class Factory:
             # decode.__class__.__call__ = Profiler(
             #     decode.__call__, out_name=cls.profile)
             decode.__class__.__call__ = TorchProfiler(
-                decode.__call__, device=cls.profile_device, out_name=cls.profile)
+                decode.__call__, out_name=cls.profile)
 
         return Multi(decoders)
