+++ openpifpaf/profiler.py
@@ -37,23 +37,19 @@ class Profiler:
 
 class TorchProfiler:
     trace_counter = 0
+    table_sort_by = 'self_cpu_time_total'
 
-    def __init__(self, function_to_profile, *, device, out_name=None):
+    def __init__(self, function_to_profile, *, out_name=None):
         if not out_name:
             out_name = 'torchprofiler_chrome_trace.json'
 
         self.function_to_profile = function_to_profile
-        self.device = device
-
         self.out_name = out_name
 
     def __call__(self, *args, **kwargs):
-        with torch.profiler.profile(activities=[
-            torch.profiler.ProfilerActivity.CPU,
-            # use_cuda=str(self.device) == 'cuda'
-        ], with_stack=True) as prof:
+        with torch.profiler.profile(with_stack=True) as prof:
             result = self.function_to_profile(*args, **kwargs)
-        print(prof.key_averages())
+        print(prof.key_averages().table(sort_by=self.table_sort_by, row_limit=-1))
 
         self.__class__.trace_counter += 1
         tracefilename = '{}.{}.json'.format(
